requirements.txt
flask==3.0.0
anthropic>=0.40.0
requests==2.31.0
apscheduler==3.10.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
twilio>=9.0.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
